+++ user/UserBase.h
@@ -7,7 +7,7 @@ class UserBase {
 public:
   void addUser(const User& user);
   std::set<User> getUsers() const;
-  int getUsersAmount() const;
+  size_t getUsersAmount() const;
   std::set<User>::iterator getUserByName(const string& name) const;
   std::set<User>::iterator getUserByLogin(const string& login) const;
 private:
